% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrees.R
\name{addTrees}
\alias{addTrees}
\title{Add Trees}
\usage{
addTrees(forest, numTreesToAdd, savePath = NULL,
  savePath.overwrite = c("warn", "delete", "merge"),
  forest.output = c("online", "offline"), cores = getCores(),
  displayProgress = TRUE)
}
\arguments{
\item{forest}{An existing forest.}

\item{numTreesToAdd}{The number of trees to add.}

\item{savePath}{If saving the forest, the directory to save to. Default is
\code{NULL}. Note that you need to re-specify the path if you're modifying
a previously saved forest.}

\item{savePath.overwrite}{If \code{savePath} is pointing to an existing
directory, possibly containing another forest, this specifies what should
be done.}

\item{forest.output}{This parameter only applies if \code{savePath} has been
set; set to 'online' (default) and the saved forest will be loaded into
memory after being trained. Set to 'offline' and the forest is not saved
into memory, but can still be used in a memory unintensive manner.}

\item{cores}{The number of cores to be used for training the new trees.}

\item{displayProgress}{A logical indicating whether the progress should be
displayed to console; default is \code{TRUE}. Useful to set to FALSE in
some automated situations.}
}
\value{
A new forest with the original and additional trees.
}
\description{
Add more trees to an existing forest. Most parameters are extracted from the
previous forest.
}
