% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_predictions.R
\name{CompetingRiskPredictions}
\alias{CompetingRiskPredictions}
\alias{extractCIF}
\alias{extractCHF}
\alias{extractSurvivorCurve}
\alias{extractMortalities}
\title{Competing Risk Predictions}
\usage{
extractCIF(x, event)

extractCHF(x, event)

extractSurvivorCurve(x)

extractMortalities(x, event, time)
}
\arguments{
\item{x}{The predictions output from a competing risk random forest.}

\item{event}{The event who's CIF/CHF/Mortality you are interested in.}

\item{time}{The time to evaluate the mortality for (relevant only for
\code{extractMortalities}).}
}
\description{
\code{extractCIF} extracts the cumulative incidence function for a prediction.

\code{extractCHF} extracts the cause-specific cumulative hazard function for a prediction.

\code{extractSurvivorCurve} extracts the Kaplan-Meier estimator
of the overall survivor curve for a prediction.

\code{extractMortalities} extracts the cause-specific
mortalities for a function, which here is the CIF integrated from 0 to
\code{time}.
}
