% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertRListToJava}
\alias{convertRListToJava}
\title{convertRListToJava}
\usage{
convertRListToJava(lst)
}
\arguments{
\item{lst}{The R list containing rJava objects}
}
\value{
An rJava List object to be used internally.
}
\description{
An internal function that converts an R list of rJava objects into a
java.util.List rJava object containing those objects. It's used internally,
and is only available because it's used in some examples that demonstrate what
other objects do.
}
\examples{
x <- Numeric(1:5)
class(x)
x <- convertRListToJava(x)
class(x)
}
\keyword{internal}
