% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToOnlineForest.R
\name{convertToOnlineForest}
\alias{convertToOnlineForest}
\title{Convert to Online Forest}
\usage{
convertToOnlineForest(forest)
}
\arguments{
\item{forest}{The offline forest.}
}
\value{
An online, in memory forst.
}
\description{
Some forests are too large to store in memory and have been saved to disk.
They can still be used, but their performance is much slower. If there's
enough memory, they can easily be converted into an in-memory forest that is
faster to use.
}
