% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_components.R
\name{CR_ResponseCombiner}
\alias{CR_ResponseCombiner}
\title{Competing Risk Response Combiner}
\usage{
CR_ResponseCombiner(events)
}
\arguments{
\item{events}{A vector of integers specifying which competing risk events's
functions should be processed. This should correspond to all of the
competing risk events that can occur, from 1 to the largest number.}
}
\value{
A response combiner object to be used in \code{\link{train}}; not
  useful on its own. However, internally, a response combiner object is a
  list consisting of the following objects: \describe{
  \item{\code{javaObject}}{The java object used in the algorithm}
  \item{\code{call}}{The call (used in \code{print})}
  \item{\code{outputClass}}{The R class of the outputs; used in
  \code{\link{predict.JRandomForest}}} \item{\code{convertToRFunction}}{An R
  function that converts a Java prediction from the combiner into R output
  that is readable by a user.} }
}
\description{
Creates a CompetingRiskResponseCombiner rJava object, which is used
internally for constructing a forest. It is used when each tree in the forest
is constructed, as it combines response level information (u & delta) into
functions such as cumulative incidence curves, cause-specific cumulative
hazard functions, and an overall Kaplan-Meier curve. This combination is done
for each terminal node for each tree.
}
\details{
The user only needs to pass this object into \code{\link{train}} as the
\code{nodeResponseCombiner} parameter.
}
\examples{
T1 <- rexp(1000)
T2 <- rweibull(1000, 1, 2)
C <- rexp(1000)

u <- round(pmin(T1, T2, C))
# ...

forestCombiner <- CR_ResponseCombiner(1:2) # there are two possible events
}
