% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric.R
\name{Numeric}
\alias{Numeric}
\title{Numeric}
\usage{
Numeric(y)
}
\arguments{
\item{y}{The R vector of numbers}
}
\value{
An R list containing rJava Doubles.
}
\description{
An internal function that converts an R vector of numerics or integers into an R list containing java.lang.Double objects. This method does not need to be used directly by the user, as \code{\link{train}} will automatically handle numeric responses if you're working in the regression settings.
}
\examples{
x <- Numeric(1:5)
class(x[[1]])
}
\keyword{internal}
