% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionComponents.R
\name{WeightedVarianceSplitFinder}
\alias{WeightedVarianceSplitFinder}
\title{WeightedVarianceSplitFinder}
\usage{
WeightedVarianceSplitFinder()
}
\value{
A split finder object to be used in \code{\link{train}}; not
  useful on its own.
}
\description{
This split finder is used in regression random forests. When a split is made,
this finder computes the sample variance in each group (divided by n, not
n-1); it then minimizes the the sum of these variances, each of them weighted
by their sample size divided by the total sample size of that node.
}
\note{
There are other split finders that are used in regression random
  forests that are not included in this package. This package is oriented
  toward the competing risk side of survival analysis; the regression options
  are provided as an example of how extensible the back-end Java package is.
  If you are interested in using this package for regression (or other uses),
  feel free to write your own components. It's really not hard to write these
  components; the WeightedVarianceSplitFinder Java class is quite short; most
  of the code is to reuse calculations from previous considered splits.
}
\examples{
splitFinder <- WeightedVarianceSplitFinder()
# You would then use it in train()

}
\references{
https://kogalur.github.io/randomForestSRC/theory.html#section8.3
}
