% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wihs.R
\docType{data}
\name{wihs}
\alias{wihs}
\title{Women's Interagency HIV Study}
\format{A data frame with 1164 rows and 6 variables: \describe{
  \item{time}{time to the event} \item{status}{denotes which event occurred.
  0 denotes censoring, 1 denotes HIV treatment began, and 2 denotes AIDS
  developed or the patient died} \item{ageatfda}{patient age at time first
  treatment approved} \item{idu}{binary specifying if the patient has a
  history of drug injections (1 if true)} \item{black}{binary specifying if
  the patient is black (1 if true)} \item{cd4nadir}{blood count of CD4 cells}
  }}
\source{
The data was obtained from the randomForestSRC R package.
}
\usage{
wihs
}
\description{
A dataset containing competing risks information for women with HIV;
recording the time to treatment, or the time to developing AIDS or death. The
time may also be censored.
}
\references{
Bacon MC, von Wyl V, Alden C, Sharp G, Robison E, Hessol N, Gange
S, Barranday Y, Holman S, Weber K, Young MA (2005). “The Women’s Interagency
HIV Study: an Observational Cohort Brings Clinical Sciences to the Bench.”
Clinical and Vaccine Immunology, 12(9), 1013–1019.
doi:10.1128/CDLI.12.9.1013-1019.2005.

Ishwaran, H., and Udaya B. Kogalur. 2018. Random Forests for Survival,
Regression, and Classification (Rf-Src). manual.
https://cran.r-project.org/package=randomForestSRC.
}
\keyword{datasets}
