% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionComponents.R
\name{MeanResponseCombiner}
\alias{MeanResponseCombiner}
\title{MeanResponseCombiner}
\usage{
MeanResponseCombiner()
}
\description{
This response combiner is used in regression random forests, where the
response in the data is a single number that needs to be averaged in each
terminal node, and then averaged across trees. This response combiner is
appropriate as an argument for both the \code{nodeResponseCombiner} and
\code{forestResponseCombiner} parameters in \code{\link{train}} when doing
regression.
}
\examples{
responseCombiner <- MeanResponseCombiner()
# You would then use it in train()

}
